chore: rename primary command signin -> login
