+++ packages/cli/src/commands/logout.rs
